modal/requirements.txt
modal
pydantic
Pillow
pytesseract
fastapi